# Improve Queue#clean performance