--- a/lib/queue.js
+++ b/lib/queue.js
@@ -797,8 +797,6 @@ Queue.prototype.clean = function (grace, type) {
   var _this = this;
 
   return new Promise(function (resolve, reject) {
-    var getter;
-
     if(grace === undefined || grace === null) {
       return reject(new Error('You must define a grace period.'));
     }
@@ -809,26 +807,14 @@ Queue.prototype.clean = function (grace, type) {
 
     if(_.indexOf([
       'completed',
-      'waiting',
+      'wait',
       'active',
       'delayed',
       'failed'], type) === -1){
       return reject(new Error('Cannot clean unkown queue type'));
     }
 
-    getter = 'get' + type.charAt(0).toUpperCase() + type.slice(1);
-
-    _this[getter]().then(function (jobs) {
-      // take all jobs outside of the grace period
-      return Promise.filter(jobs, function (job) {
-        return job && ((!job.timestamp) || (job.timestamp < Date.now() - grace));
-      });
-    }).then(function (jobs) {
-      // remove those old jobs
-      return Promise.each(jobs, function (job) {
-        return job.remove(_this.token);
-      });
-    }).then(function (jobs) {
+    return scripts.cleanJobsInSet(_this, type, Date.now() - grace).then(function (jobs) {
       _this.emit('cleaned', jobs, type);
       resolve(jobs);
       return null;