--- a/lib/queue.js
+++ b/lib/queue.js
@@ -791,9 +791,10 @@ Queue.prototype.toKey = function(queueType){
  *
  * @param {int} grace - The grace period
  * @param {string} [type=completed] - The type of job to clean. Possible values
+ * @param {int} The max number of jobs to clean
  * are completed, waiting, active, delayed, failed. Defaults to completed.
  */
-Queue.prototype.clean = function (grace, type) {
+Queue.prototype.clean = function (grace, type, limit) {
   var _this = this;
 
   return new Promise(function (resolve, reject) {
@@ -814,7 +815,7 @@ Queue.prototype.clean = function (grace, type) {
       return reject(new Error('Cannot clean unkown queue type'));
     }
 
-    return scripts.cleanJobsInSet(_this, type, Date.now() - grace).then(function (jobs) {
+    return scripts.cleanJobsInSet(_this, type, Date.now() - grace, limit).then(function (jobs) {
       _this.emit('cleaned', jobs, type);
       resolve(jobs);
       return null;