--- a/lib/scripts.js
+++ b/lib/scripts.js
@@ -374,21 +374,25 @@ var scripts = {
 
   cleanJobsInSet: function(queue, set, ts) {
     var command;
+    var removeCommand;
     var hash;
     switch(set) {
       case 'completed':
       case 'failed':
-        command = 'local JOBS = redis.call("SMEMBERS", KEYS[1])';
+        command = 'local jobs = redis.call("SMEMBERS", KEYS[1])';
+        removeCommand = 'redis.call("SREM", KEYS[1], job)';
         hash = 'cleanSet';
         break;
       case 'wait':
       case 'active':
       case 'paused':
-        command = 'local JOBS = redis.call("LRANGE", KEYS[1], 0, -1)';
+        command = 'local jobs = redis.call("LRANGE", KEYS[1], 0, -1)';
+        removeCommand = 'redis.call("LREM", KEYS[1], 0, job)';
         hash = 'cleanList';
         break;
       case 'delayed':
-        command = 'local JOBS = redis.call("ZRANGE", KEYS[1], 0, -1)';
+        command = 'local jobs = redis.call("ZRANGE", KEYS[1], 0, -1)';
+        removeCommand = 'redis.call("ZREM", KEYS[1], job)';
         hash = 'cleanOSet';
         break;
     }
@@ -397,17 +401,13 @@ var scripts = {
       command,
       'local deleted = {}',
       'local jobTS',
-      'for _, job in ipairs(JOBS) do',
-      '  if (redis.call("EXISTS", ARGV[1] .. job ..  ":lock") == 0) then',
-      '    jobTS = redis.call("HGET", ARGV[1] .. job, "timestamp")',
+      'for _, job in ipairs(jobs) do',
+      '  local jobKey = ARGV[1] .. job',
+      '  if (redis.call("EXISTS", jobKey ..  ":lock") == 0) then',
+      '    jobTS = redis.call("HGET", jobKey, "timestamp")',
       '    if(not jobTS or jobTS < ARGV[2]) then',
-      '      redis.call("LREM", ARGV[1] .. ARGV[3], 0, job)',
-      '      redis.call("LREM", ARGV[1] .. ARGV[4], 0, job)',
-      '      redis.call("ZREM", ARGV[1] .. ARGV[5], job)',
-      '      redis.call("LREM", ARGV[1] .. ARGV[6], 0, job)',
-      '      redis.call("SREM", ARGV[1] .. ARGV[7], job)',
-      '      redis.call("SREM", ARGV[1] .. ARGV[8], job)',
-      '      redis.call("DEL", ARGV[1] .. job)',
+      removeCommand,
+      '      redis.call("DEL", jobKey)',
       '      table.insert(deleted, job)',
       '    end',
       '  end',
@@ -422,13 +422,7 @@ var scripts = {
       1,
       queue.toKey(set),
       queue.toKey(''),
-      ts,
-      'active',
-      'wait',
-      'delayed',
-      'paused',
-      'completed',
-      'failed',
+      ts
     ];
 
     return execScript.apply(scripts, args);