--- a/lib/scripts.js
+++ b/lib/scripts.js
@@ -372,10 +372,13 @@ var scripts = {
     return execScript.apply(scripts, args);
   },
 
-  cleanJobsInSet: function(queue, set, ts) {
+  cleanJobsInSet: function(queue, set, ts, limit) {
     var command;
     var removeCommand;
+    var breakEarlyCommand = '';
     var hash;
+    limit = limit || 0;
+
     switch(set) {
       case 'completed':
       case 'failed':
@@ -397,17 +400,31 @@ var scripts = {
         break;
     }
 
+    if(limit > 0) {
+      breakEarlyCommand = [
+        'if deletedCount >= limit then',
+        '  break',
+        'end',
+      ].join('\n');
+
+      hash = hash + 'WithLimit';
+    }
+
     var script = [
       command,
       'local deleted = {}',
+      'local deletedCount = 0',
+      'local limit = tonumber(ARGV[3])',
       'local jobTS',
       'for _, job in ipairs(jobs) do',
+      breakEarlyCommand,
       '  local jobKey = ARGV[1] .. job',
       '  if (redis.call("EXISTS", jobKey ..  ":lock") == 0) then',
       '    jobTS = redis.call("HGET", jobKey, "timestamp")',
       '    if(not jobTS or jobTS < ARGV[2]) then',
       removeCommand,
       '      redis.call("DEL", jobKey)',
+      '      deletedCount = deletedCount + 1',
       '      table.insert(deleted, job)',
       '    end',
       '  end',
@@ -422,7 +439,8 @@ var scripts = {
       1,
       queue.toKey(set),
       queue.toKey(''),
-      ts
+      ts,
+      limit
     ];
 
     return execScript.apply(scripts, args);