@manast 2016-06-18
this is a great PR. Just one important consideration. If deleting the jobs takes 3 seconds, it will mean that redis will be blocked for 3 seconds, since it is performed atomically and no other operations can be performed until it has finished. This can be acceptable or not, depending on the use case.