@chuym 2016-06-20
> If deleting the jobs takes 3 seconds, it will mean that redis will be blocked for 3 seconds

Yeah that's a downside of this. Another idea is to batch the deletes in `1000` batches perhaps, to diminish the time this script blocks. Thoughts?