@manast 2016-06-20
@chuym I have given you some hints on how to optimize the script, please read my latest comments in the code.
The ultimate solution would be to make this function iterative, as you mentioned, in 1000 item batches, which would be resolved in a millisecond or so.