@chuym 2016-06-20
@manast I've ammended the PR, please take a second look when you have the time. Thanks for following up!

After pondering a bit, I decided not to remove in batches since it can be problematic: If we delete in batches, how do we know when we are done? I added support to optionally limit the number of jobs to clean by passing an optional paramter to `Queue#clean`. This way we give clients the tools to handle that themselves.